:sparkles:Add action edit and create new owner, also better error message

+++ VetiCare_Frontend/src/pages/duenosAdmin.jsx
@@ -7,6 +7,7 @@ import Modal from "../components/admin-edit-modal"
 import EditAppointmentForm from "../components/edit-appointment"
 import EditDuenoForm from "../components/edit-duenos"
 import AddButton from "../components/add-button"
+import AddOwnerForm from "../components/add-admin-owner-form"
 
 import Swal from 'sweetalert2';
 import "sweetalert2/dist/sweetalert2.min.css";
@@ -16,11 +17,27 @@ function DuenosAdmin() {
 
   const token = localStorage.getItem("token");
   const API_URL = import.meta.env.VITE_API_URL;
+  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
+  const [isAddModalOpen, setIsAddModalOpen] = useState(false);
+
   const [isModalOpen, setModalOpen] = useState(false);
   const [ownerToEdit, setOwnerToEdit] = useState(null);
   const [owners, setOwners] = useState([]);
   const [filteredOwners, setFilteredOwners] = useState([]);
   const [error, setError] = useState(null);
+  const [isCreating, setIsCreating] = useState(false);
+
+  // SweetAlert arriba
+  useEffect(() => {
+    const style = document.createElement("style");
+    style.textContent = `
+      .swal2-container, .swal2-popup {
+        z-index: 999999 !important;
+      }
+    `;
+    document.head.appendChild(style);
+    return () => document.head.removeChild(style);
+  }, []);
 
 
   const adminOwnerColumns = [
@@ -34,7 +51,8 @@ function DuenosAdmin() {
     const currentPage = 1;
     const itemsPerPage = 7;
     try {
-      const response = await fetch(`${API_URL}/api/users/owners`, {
+      //const response = await fetch(`${API_URL}/api/users/owners`, {
+      const response = await fetch(`${API_URL}/api/users`, {
         method: 'GET',
         headers: {
           Authorization: `Bearer ${token}`,
@@ -49,8 +67,10 @@ function DuenosAdmin() {
         });
       }
       const data = await response.json();
-      console.log(data)
-      const filteredData = data.map((item) => ({
+
+      const ownersOnly = data.filter(u => u.role_id === 1);
+
+      const filteredData = ownersOnly.map((item) => ({
         id: item.id,
         rowNumber: (currentPage - 1) * itemsPerPage + index + 1,
         DUI: item.dui,
@@ -114,6 +134,13 @@ function DuenosAdmin() {
     }
   }
 
+  function handleCreate() {
+    //setIsCreating(true);
+    setOwnerToEdit(null);
+    setModalOpen(true);
+  }
+
+
   const handleEdit = async (ownerId) => {
     if (!ownerId) {
       setOwnerToEdit(null);
@@ -135,6 +162,28 @@ function DuenosAdmin() {
 
   async function deleteOwner(id) {
     try {
+
+      const owner = await getById(id);
+      const currentStatus = owner.status_id;
+      const newStatus = currentStatus === 1 ? 2 : 1;
+
+      const actionText = newStatus === 2 ? "desactivar" : "activar";
+      const actionTextPast = newStatus === 2 ? "desactivado" : "activado";
+
+      const confirm = await Swal.fire({
+        icon: "warning",
+        title: `¿Desea ${actionText} este usuario?`,
+        text: `El usuario será ${actionTextPast}. Esta acción es reversible.`,
+        showCancelButton: true,
+        confirmButtonColor: "rgba(161, 19, 19, 1)",
+        cancelButtonColor: "#374f59",
+        confirmButtonText: `Sí, ${actionText}`,
+        cancelButtonText: "Cancelar",
+      });
+
+      if (!confirm.isConfirmed) return;
+
+
       const response = await fetch(`${API_URL}/api/users/${id}`, {
         method: 'DELETE',
         headers: {
@@ -160,6 +209,33 @@ function DuenosAdmin() {
       });
     }
   }
+
+  function mensajeErrorForm(errorText) {
+    errorText = errorText.toLowerCase();
+
+    if (errorText.includes("duplicate") || errorText.includes("already in use")) {
+      return "El correo o DUI ingresado ya están en uso.";
+    }
+
+    if (errorText.includes("duiformat")) {
+      return "El DUI ingresado no tiene el formato correcto, por favor agregar guion. Ejemplo: 12345678-9";
+    }
+
+    if (errorText.includes("email")) {
+      return "Correo electrónico ingresado no es válido";
+    }
+
+    if (errorText.includes("phone") || errorText.includes("telefono")) {
+      return "El teléfono debe tener el formato ####-####, incluyendo el guion.";
+    }
+
+    if (errorText.includes("full_name") || errorText.includes("name")) {
+      return "El nombre no puede contener caracteres especiales ni estar vacío.";
+    }
+
+    return "Ocurrió un error al procesar la solicitud. Verifique los datos ingresados.";
+  }
+
 
   async function updateOwner(id, updatedData) {
     if (id == null) {
@@ -192,7 +268,7 @@ function DuenosAdmin() {
           });
         }
         const result = await response.json();
-        console.log('Owner added:', result);
+        //console.log('Owner added:', result);
         setModalOpen(false);
         getData();
       } catch (error) {
@@ -223,11 +299,17 @@ function DuenosAdmin() {
           }),
         });
         if (!response.ok) {
-          const errorText = await response.text();
-          console.error(`Error ${response.status}:`, errorText);
-          const parsedError= errorText.match(/failed on the '([^']+)' tag/);
-          alert(`Error al agregar nuevo dueño: ${parsedError}`);
-          throw new Error(`Error ${response.status}: Failed to update appointment`);
+          const rawError = await response.text();
+          const friendlyMessage = mensajeErrorForm(rawError);
+
+          Swal.fire({
+            icon: "error",
+            title: "Error al actualizar",
+            text: friendlyMessage,
+            confirmButtonText: "Entendido"
+          });
+
+          return;
         }
         const result = await response.json();
         Swal.fire({
@@ -247,6 +329,50 @@ function DuenosAdmin() {
     }
   }
 
+  async function addOwner(newOwner) {
+    try {
+      const response = await fetch(`${API_URL}/api/users/register`, {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+          "Authorization": `Bearer ${token}`,
+        },
+        body: JSON.stringify({
+          full_name: newOwner.full_name,
+          dui: newOwner.dui,
+          phone: newOwner.phone,
+          email: newOwner.email,
+          password_hash: newOwner.password_hash,
+          role_id: 1,
+          status_id: 1
+        }),
+      });
+
+      if (!response.ok) {
+        const rawError = await response.text();
+        Swal.fire({
+          icon: "error",
+          title: "No se pudo crear el dueño",
+          text: rawError,
+        });
+        return;
+      }
+
+      Swal.fire({
+        icon: "success",
+        title: "Dueño creado",
+        text: "Registro completado con éxito",
+        timer: 3500,
+        showConfirmButton: true
+      });
+
+      getData();
+      setModalOpen(false);
+
+    } catch (error) {
+      Swal.fire("Error", "No se pudo conectar con el servidor", "error");
+    }
+  }
 
 
 
@@ -272,12 +398,17 @@ function DuenosAdmin() {
         }}>Dueños</h2>
         <div className="search-add-row">
           <SearchBox onSearch={handleSearch} placeholder="Busque dueño por nombre" />
-          <AddButton onClick={handleEdit} />
+          <AddButton onClick={handleCreate} />
         </div>
-        <AdminTable rows={filteredOwners} columns={adminOwnerColumns} onEdit={handleEdit} onDelete={deleteOwner}/>
+        <AdminTable rows={filteredOwners} columns={adminOwnerColumns} onEdit={handleEdit} onDelete={deleteOwner} />
         <Modal isOpen={isModalOpen} onClose={() => setModalOpen(false)}>
-        <EditDuenoForm initialData={ownerToEdit} onSubmit={updateOwner} />
+          {ownerToEdit ? (
+            <EditDuenoForm initialData={ownerToEdit} onSubmit={updateOwner} />
+          ) : (
+            <AddOwnerForm onSubmit={addOwner} />
+          )}
         </Modal>
+
       </div>
     </Layout>
   )
